+++ popin/community/urls.py
@@ -13,6 +13,9 @@ urlpatterns = [
     # path('write/status/', views.write_status, name='write_status'),
     #
     path('chgReview/', views.chgReview, name='chgReview'),
+    path('chgReview/main/', views.chgReviewmain, name='chgReviewmain'),
+    path('chgReview/view/', views.chgReviewview, name='chgReviewview'),
+    
     path('companion/', views.companion, name='companion'),
     path('proxy/', views.proxy, name='proxy'),
     path('sharing/', views.sharing, name='sharing'),
